books/requirements.txt
asgiref==3.5.2
certifi==2022.5.18.1
cffi==1.15.0
charset-normalizer==2.0.12
coverage==6.4.1
cryptography==37.0.2
defusedxml==0.7.1
Django==3.2.13
django-filter==21.1
djangorestframework==3.13.1
idna==3.3
oauthlib==3.2.0
psycopg2==2.9.3
pycparser==2.21
PyJWT==2.4.0
python3-openid==3.2.0
pytz==2022.1
requests==2.28.0
requests-oauthlib==1.3.1
social-auth-app-django==5.0.0
social-auth-core==4.2.0
sqlparse==0.4.2
urllib3==1.26.9
